requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.5.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3